major overhaul, change all numbers to numeric(12, 3) from of int, modify newDocument functions and references to fix type errors, modify queries

+++ sql/schema/002_purchases.sql
@@ -4,8 +4,8 @@ CREATE TABLE purchases (
     created_at TIMESTAMP NOT NULL,
     updated_at TIMESTAMP NOT NULL,
     suplier TEXT NOT NULL,
-    price INT NOT NULL,
-    quantity INT NOT NULL,
+    price NUMERIC(12, 3) NOT NULL,
+    quantity NUMERIC(12, 3) NOT NULL,
     grain_type TEXT NOT NULL
 );
